Code refactored and bug fixed

+++ server/src/api/services/events.service.js
@@ -29,11 +29,12 @@ export const getUserEvents = async (userId) => {
 /*
  * Create a new Event
  */
-export const createEvent = async (event) => {
+export const createEvent = async (event, creator) => {
   return new Promise(async (resolve, reject) => {
     // Create a new Event
     const newEvent = new Event({
       ...event,
+      creator,
     });
 
     // Save and return the new Event
@@ -52,45 +53,27 @@ export const createEvent = async (event) => {
  * Update the Event
  *
  * */
-export const updateEvent = async (_id, edits) => {
-  return new Promise(async (resolve, reject) => {
-    // Update the event according to _id
-    const updatedEvent = await Event.findOneAndUpdate(
-      { _id },
-      {
-        ...edits,
-      },
-      { returnDocument: "after" }
-    );
-    updateEvent
-      ? resolve(updatedEvent)
-      : reject({
-          message: "Event didn't update.",
-          code: "INTERNAL_SERVER_ERROR",
-          status: 500,
-        });
-  });
+export const updateEvent = async (event, userId) => {
+  const { _id, edits } = event;
+  // Update the event according to _id and userId
+  const updatedEvent = await Event.findOneAndUpdate(
+    { _id, creator: userId },
+    {
+      ...edits,
+    },
+    { returnDocument: "after" }
+  );
+
+  return updatedEvent;
 };
 
 /*
  * Delete the Event
  *
  * */
-export const deleteEvent = async (_id) => {
-  return new Promise(async (resolve, reject) => {
-    // Delete the event according to _id
-    const { deletedCount } = await Event.deleteOne({ _id });
+export const deleteEvent = async (_id, userId) => {
+  // Delete the event according to _id
+  const { deletedCount } = await Event.deleteOne({ _id, creator: userId });
 
-    // Get all events
-    const events = await Event.find().lean();
-
-    // If deletedCount is equal to 0 return null
-    deletedCount
-      ? resolve(events)
-      : reject({
-          message: "Event didn't delete",
-          code: "INTERNAL_SERVER_ERROR",
-          status: 500,
-        });
-  });
+  return deletedCount;
 };
